refactor: lsp -> internal/lsp
