+++ internal/lsp/completion.go
