+++ internal/lsp/document.go
