+++ internal/lsp/hover.go
