+++ internal/lsp/hover.json
