+++ internal/lsp/markdown/completion/configDeclare.md
