+++ internal/lsp/markdown/completion/enumType.md
