+++ internal/lsp/markdown/completion/fsmDeclare.md
