+++ internal/lsp/markdown/completion/secretDeclare.md
